% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyPcmMotifs.R
\name{classifyPcmMotifs}
\alias{classifyPcmMotifs}
\title{Classifies motifs within given PWM file}
\usage{
classifyPcmMotifs(transfacFilePath)
}
\arguments{
\item{transfacFilePath}{A string indicating the path to the TRANSFAC file}
}
\description{
A function that finds matches for given motif PWM to motifs within the MotifDb database
}
\examples{
transfacFilePath <- system.file("extdata", "new0007.txt", package = "MotifFunc")
matchNames <- classifyPcmMotifs(transfacFilePath)

PWMfile <- system.file("extdata", "MA0007.1.transfac", package = "MotifFunc")
jaspar.scores <- MotifFunc::jaspar.scores
matchNames <- MotifFunc::classifyPcmMotifs(PWMfile)

}
