% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifySeqMotifs.R
\name{classifySeqMotifs}
\alias{classifySeqMotifs}
\title{Classifies motifs within given DNA sequence}
\usage{
classifySeqMotifs(consensusSeq)
}
\arguments{
\item{consensusSeq}{A string indicating a desired consensus sequence (i.e. Contains only "A", "T", "C", "G")}
}
\description{
A function that creates a motif PWM based on given DNA consensus sequence and finds matches within the MotifDb database
}
\examples{
jaspar.scores <- MotifFunc::jaspar.scores
matchNames <- MotifFunc::classifySeqMotifs("AGCGTAGGCGT")

}
