% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFullOrganism.R
\name{getFullOrganism}
\alias{getFullOrganism}
\title{Gets full species name of organism}
\usage{
getFullOrganism(MotifDbOrganism)
}
\arguments{
\item{MotifDbOrganism}{A string indicating organism/species names (based on MotifDb records)}
}
\value{
Returns fullName - A string indicating full species name
}
\description{
A function that takes organism names in MotifDb format, retrieves corresponding full species name
}
\examples{
getFullOrganism("Hsapiens")


}
