% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctionWC.R
\name{getFunctionWC}
\alias{getFunctionWC}
\title{Creates word cloud visualization of GO functionality based on Motif match
names from MotifDb}
\usage{
getFunctionWC(matchNames)
}
\arguments{
\item{matchNames}{A character vector indicating Motif "names" (based on
MotifDb records)}
}
\value{
Returns functionFreq - A table/list indicating frequency of found
functionalities
}
\description{
A function that takes motif "names" in MotifDb format, retrieves
corresponding GO functions and produces a word cloud visualization based on
function frequency within all specified motifs.
}
\examples{
matchNames <- MotifFunc::classifySeqMotifs("AGCGTAGGCGT")
functionFreq <- MotifFunc::getFunctionWC(matchNames)

transfacFilePath <- system.file("extdata", "new0007.txt", package = "MotifFunc")
matchNames <- MotifFunc::classifyPcmMotifs(transfacFilePath)
functionFreq <- MotifFunc::getFunctionWC(matchNames)

}
\references{
Hines, K. (2014). Stack Overflow source code [Source code].
https://stackoverflow.com/questions/26937960/creating-word-cloud-of-phrases-not-individual-words-in-r.
}
