% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jaspar.scores}
\alias{jaspar.scores}
\title{Jaspar.scores for MotIV::motifMatch()}
\format{An double containing jsapar scores from jaspar 2010, taken from MotIV documentation}
\source{
\url{http://bioconductor.org/packages/release/bioc/vignettes/MotIV/inst/doc/MotIV.pdf}
}
\usage{
jaspar.scores
}
\description{
Jaspar.scores for MotIV::motifMatch()
}
\examples{
\dontrun{
 jaspar.scores
}
}
\keyword{datasets}
