% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matches}
\alias{matches}
\title{Match data produced by MotIV::motifMatch()}
\format{An S4 of a motiv object typically outputted by MotIV::motifMatch()}
\source{
\url{http://www.disgenet.org/static/disgenet_ap1/files/downloads/all_gene_disease_pmid_associations.tsv.gz}
}
\usage{
matches
}
\description{
Match data produced by MotIV::motifMatch()
}
\examples{
\dontrun{
 matches
}
}
\keyword{datasets}
