% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotIV.toTable.R
\name{MotIV.toTable}
\alias{MotIV.toTable}
\title{Converts S4 output of MotIV::motifMatch() to a list}
\usage{
MotIV.toTable(match)
}
\arguments{
\item{match}{An S4 typically outputted by MotIV::motifMatch()}
}
\value{
Returns df - A list indicating motif information as outlined in MotDb
}
\description{
A function that takes S4 outputted by MotIV::motifMatch() and returns a list containing match information
}
\examples{
data("matches")
MotIV.toTable("Hsapiens")

}
