% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{organismFullNames}
\alias{organismFullNames}
\title{Full species names mapped out to the appreviated ones given by MotifDb}
\format{}
\usage{
organismFullNames
}
\description{
Abbreviated organisms were taken from those stored in MotifDb that are
associated with motifs.
There is sometimes more than 1 organism associated to a motif.
}
\examples{
\dontrun{
 organismFullNames
}
}
\keyword{datasets}
