% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{pickOrganism}
\alias{pickOrganism}
\title{Retrieves user input for organism selection}
\usage{
pickOrganism(organismVec)
}
\arguments{
\item{organismVec}{An character vector containing species names}
}
\value{
Returns a string indicating chosen organism
}
\description{
A function that displays organism options to user and returns chosen organism
}
\examples{
organismVec <- c("Mus musculus", "Saccharomyces cerevisiae", "Homo sapiens")
pickOrganism(organismVec)

}
