% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMotifFunc.R
\name{runMotifFunc}
\alias{runMotifFunc}
\title{Launch the shiny app for package MotifFunc}
\usage{
runMotifFunc()
}
\value{
No return value but open up a shiny page.
}
\description{
A function that launches the shiny app for this package.
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
runMotifFunc()
}

}
