% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctionWC.R
\name{getFunctionWC}
\alias{getFunctionWC}
\title{Creates word cloud visualization of GO functionality based on Motif match names from MotifDb}
\usage{
getFunctionWC(match_names)
}
\arguments{
\item{match_names}{A character vector indicating Motif "names" (based on MotifDb records)}
}
\description{
A function that takes motif "names" in MotifDb format, retrieves corresponding GO functions and produces a word cloud visualization based on function frequency within all specified motifs.
}
\examples{
match_names <- classifySeqMotifs("AGCGTAGGCGT")
functionFreq <- getFunctionWC(match_names)

match_names <- classifyPcmMotifs("/Users/minhanho/Documents/MotifFunc/new0007.txt")
functionFreq <- getFunctionWC(match_names)

}
