% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{MotIV.toTable}
\alias{MotIV.toTable}
\title{Converts S4 output of MotIV::motifMatch() to a list}
\usage{
MotIV.toTable(match)
}
\arguments{
\item{match}{An S4 typically outputted by MotIV::motifMatch()}
}
\value{
Returns df - A list indicating motif information as outlined in MotDb
}
\description{
A function that takes S4 outputted by MotIV::motifMatch() and returns a list
containing match information
}
\examples{
data("matches")
MotIV.toTable(matches)

}
\references{
Shannon P, Richards M (2019). MotifDb: An Annotated Collection of Protein-DNA
Binding Sequence Motifs. R package version 1.26.0.
\url{https://rdrr.io/bioc/MotifDb/src/inst/doc/MotifDb.R}
}
