% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{organismFullNames}
\alias{organismFullNames}
\title{Full species names mapped out to the appreviated ones given by MotifDb}
\format{An character vector, in which each element has the format:
\describe{
 \item{abreviated name;<another abbreviated name>;...}{The semi colon is only used as a seperator for multiorganism motifs}
 \item{=}{Beginning of full names}
 \item{full name=<another full name>=...}{The equals sign is only used as a seperator for multiorganism motifs}
}}
\usage{
organismFullNames
}
\description{
Abbreviated organisms were taken from those stored in MotifDb that are
associated with motifs.
There is sometimes more than 1 organism associated to a motif.
}
\examples{
\dontrun{
 organismFullNames
}
}
\keyword{datasets}
